Impl navigate to BottomAppBarType screen
